Move VR logic to main Game component to avoid code duplication
