+++ src/controls/index.ts
@@ -2,11 +2,16 @@ import {makeFirstPersonMouseControls} from './mouse';
 import {makeKeyboardControls} from './keyboard';
 import {makeFirstPersonTouchControls} from './touch';
 import { getParams } from '../params';
+import { VRControls } from './vr';
 
-export function createControls(game, elem, sceneManager) {
+export function createControls(vr, game, elem, sceneManager, renderer) {
     const params = getParams();
     let controls;
-    if (params.mobile) {
+    if (vr) {
+        controls = [
+            new VRControls(sceneManager, game, renderer)
+        ];
+    } else if (params.mobile) {
         controls = [
             makeFirstPersonTouchControls(game),
         ];
