+++ src/controls/vr.ts
@@ -6,6 +6,7 @@ import ControllerModel from './vr/ControllerModel';
 import { createMotionController } from './vr/utils';
 import { debugProfiles, updateDebug } from './vr/debugProfiles';
 import { getControllerMappings, applyMappings, Mappings } from './vr/mappings';
+import { getParams } from '../params';
 
 // Time in ms we sample the change is position to determine controller velocity.
 const VELOCITY_UPDATE_TIME = 100;
@@ -27,12 +28,12 @@ export class VRControls {
     triggered: boolean;
     skipping: boolean;
 
-    constructor(params: any, sceneManager: any, game: any, renderer: any) {
+    constructor(sceneManager: any, game: any, renderer: any) {
         this.xr = renderer.threeRenderer.xr;
         this.ctx = {
             sceneManager,
             game,
-            params,
+            params: getParams(),
             state: {}
         };
         this.controllers = {};
@@ -43,7 +44,7 @@ export class VRControls {
         this.onInputSourcesChange = this.onInputSourcesChange.bind(this);
         this.initializeVRController(0);
         this.initializeVRController(1);
-        if (params.vrCtrlDBG) {
+        if (this.ctx.params.vrCtrlDBG) {
             debugProfiles(this);
         }
     }
